Split some tests and login before testing
now that we require user to be logged in by default tests should
also make sure they are logged in before asserting the results
